+++ src/prelude.rs
@@ -106,13 +106,13 @@ pub use crate::error::{
 };
 
 // Utility functions and traits
+pub use crate::chains::csv::{OhlcvCandle, read_ohlcv_from_zip};
 pub use crate::model::utils::ToRound;
 #[allow(deprecated)]
 pub use crate::utils::setup_logger;
 pub use crate::utils::{
-    Len, OhlcvCandle, TimeFrame,
+    Len, TimeFrame,
     others::calculate_log_returns,
-    read_ohlcv_from_zip,
     time::{convert_time_frame, get_tomorrow_formatted, get_x_days_formatted},
 };